Feat: Navigation with buttons
